part 3: no-unnecessary-type-assertion

+++ src/webgpu/api/operation/sampling/anisotropy.spec.ts
@@ -300,7 +300,7 @@ g.test('anisotropic_filter_mipmap_color')
       if (entry.expected instanceof Uint8Array) {
         // equal exactly one color
         t.expectSinglePixelIn2DTexture(colorAttachment, kColorAttachmentFormat, entry.coord, {
-          exp: entry.expected as Uint8Array,
+          exp: entry.expected,
           generateWarningOnly: t.params._generateWarningOnly,
         });
       } else {
